+++ versiontools/tests.py
@@ -22,7 +22,8 @@ from distutils.errors import DistutilsSetupError
 
 from unittest import TestCase
 
-from versiontools import Version, handle_version
+from versiontools import Version
+from versiontools.setuptools_hooks import version as handle_version 
 
 
 class VersionFormattingTests(TestCase):